requirements.txt
Flask>=3.1.0
mysql-connector-python>=9.1.0
pandas>=2.2.3
plotly>=6.0.0
python-dotenv>=1.0.1
Werkzeug>=3.1.3
requests>=2.32.3
twilio>=9.4.1
weasyprint>=62.3
